Thomas' comments

+++ tests/utils.py
@@ -4,11 +4,20 @@ import shutil
 import tempfile
 from contextlib import contextmanager
 
+try:
+    from pathlib import Path
+except ImportError:
+    from pathlib2 import Path
+
+TEST_DIR = Path(__file__).resolve().parent
+TEST_DATA_DIR = TEST_DIR / 'data'
+
 
 @contextmanager
-def setup_tempdir():
+def setup_tempdir(cleanup=True):
     temp_dir = tempfile.mkdtemp()
     try:
         yield temp_dir
     finally:
-        shutil.rmtree(temp_dir)
+        if cleanup:
+            shutil.rmtree(temp_dir)
